Don't rely on the input event in IE... ever!
Normally the input event would be reliable in IE10+, but not on contenteditable elements. However, since we've hacked in
our own support for using typeahead on a contenteditable, we run into this unfortunate corner case where typeahead
doesn't work.

Technically, we could continue to use the input event in IE10+ on non-contenteditable elements, but this is a simpler
fix that works, so #shipit